requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.121.1
fastapi-limiter==0.1.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
redis==7.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
stripe==11.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-exporter-prometheus>=0.45b0
prometheus-client>=0.20.0
